refactor: OrderTable이 TableGroup을 직접참조하는 것을 간접참조로 변경하여 순환참조 제거
