+++ src/main/java/kitchenpos/application/mapper/TableGroupMapper.java
@@ -3,7 +3,7 @@ package kitchenpos.application.mapper;
 import java.util.List;
 import java.util.stream.Collectors;
 import kitchenpos.domain.ordertable.OrderTable;
-import kitchenpos.domain.tablegroup.TableGroup;
+import kitchenpos.domain.ordertable.TableGroup;
 import kitchenpos.dto.request.TableGroupRequest;
 import kitchenpos.dto.request.TableGroupRequest.TableId;
 import kitchenpos.repository.OrderTableRepository;
