+++ src/main/java/kitchenpos/repository/JpaTableGroupRepository.java
@@ -1,6 +1,6 @@
 package kitchenpos.repository;
 
-import kitchenpos.domain.tablegroup.TableGroup;
+import kitchenpos.domain.ordertable.TableGroup;
 import org.springframework.data.repository.Repository;
 
 public interface JpaTableGroupRepository extends Repository<TableGroup, Long>, TableGroupRepository {
