+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -152,7 +152,7 @@ class TableGroupServiceTest extends ServiceTest {
             assertAll(() -> {
                 assertThat(tableGroupRepository.findById(tableGroup.getId())).isNotEmpty();
                 assertThat(orderTableRepository.findAll())
-                        .allMatch(orderTable -> Objects.isNull(orderTable.getTableGroup()))
+                        .allMatch(orderTable -> Objects.isNull(orderTable.getTableGroupId()))
                         .allMatch(orderTable -> !orderTable.isEmpty());
             });
         }
